[REFACTORING] Split IMAPServerAppendNoLimitTest
