+++ server/protocols/protocols-imap4/src/test/java/org/apache/james/imapserver/netty/IMAPServerTest.java
@@ -234,99 +234,6 @@ class IMAPServerTest {
     }
 
 
-    @Nested
-    class NoLimit {
-        IMAPServer imapServer;
-        private int port;
-
-        @BeforeEach
-        void beforeEach() throws Exception {
-            imapServer = createImapServer("imapServerNoLimits.xml");
-            port = imapServer.getListenAddresses().get(0).getPort();
-        }
-
-        @AfterEach
-        void tearDown() {
-            imapServer.destroy();
-        }
-
-        @Test
-        void smallAppendsShouldWork() throws Exception {
-            assertThatCode(() ->
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .append("INBOX", SMALL_MESSAGE))
-                .doesNotThrowAnyException();
-
-            assertThat(testIMAPClient.select("INBOX")
-                .readFirstMessage())
-                .contains("\r\n" + SMALL_MESSAGE + ")\r\n");
-        }
-
-        @Test
-        void capabilityAdvertizeAppendLimit() throws Exception {
-            assertThat(
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .capability())
-                .contains("APPENDLIMIT")
-                .doesNotContain("APPENDLIMIT=");
-        }
-
-        @Test
-        void statusAdvertizeAppendLimit() throws Exception {
-            assertThat(
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .sendCommand("STATUS \"INBOX\" (APPENDLIMIT)"))
-                .contains("* STATUS \"INBOX\" (APPENDLIMIT NIL)");
-        }
-
-        @Test
-        void mediumAppendsShouldWork() throws Exception {
-            assertThatCode(() ->
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .append("INBOX", _65K_MESSAGE))
-                .doesNotThrowAnyException();
-
-            assertThat(testIMAPClient.select("INBOX")
-                .readFirstMessage())
-                .contains("\r\n" + _65K_MESSAGE + ")\r\n");
-        }
-
-        @Test
-        void loginFixationShouldBeRejected() throws Exception {
-            InMemoryMailboxManager mailboxManager = memoryIntegrationResources.getMailboxManager();
-            mailboxManager.createMailbox(
-                MailboxPath.forUser(USER, "pwnd"),
-                mailboxManager.createSystemSession(USER));
-            mailboxManager.createMailbox(
-                MailboxPath.forUser(USER2, "notvuln"),
-                mailboxManager.createSystemSession(USER2));
-
-            testIMAPClient.connect("127.0.0.1", port)
-                // Injected by a man in the middle attacker
-                .rawLogin(USER.asString(), USER_PASS);
-
-            assertThatThrownBy(() -> testIMAPClient.rawLogin(USER2.asString(), USER_PASS))
-                .isInstanceOf(IOException.class)
-                .hasMessage("Login failed");
-        }
-
-        @RepeatedTest(200)
-        void largeAppendsShouldWork() throws Exception {
-            assertThatCode(() ->
-                testIMAPClient.connect("127.0.0.1", port)
-                    .login(USER.asString(), USER_PASS)
-                    .append("INBOX", _129K_MESSAGE))
-                .doesNotThrowAnyException();
-
-            assertThat(testIMAPClient.select("INBOX")
-                .readFirstMessage())
-                .contains("\r\n" + _129K_MESSAGE + ")\r\n");
-        }
-    }
 
     @Nested
     class AppendNonSynchronizedLitterals {
